Fix overflow of counters incremented on each I/O operation
 - In the thread_stat struct: uint32_t counters are updated to uint64_t.
 - In the io_u_plat_entry struct: unsigned int counters are updated to
   uint64_t.
It fixes overflow of these counters.

+++ io_u.c
@@ -1012,7 +1012,7 @@ out:
 	return 0;
 }
 
-static void __io_u_mark_map(unsigned int *map, unsigned int nr)
+static void __io_u_mark_map(uint64_t *map, unsigned int nr)
 {
 	int idx = 0;
